Removing unnecessary directory
