--- DatabaseAssignmentGUI/safeguardexample.py
@@ -1,4 +1,0 @@
-host="myHost",
-user="myUser",
-password="myPassword",
-database="myDatabase"
