+++ process_assignments.py
@@ -1,0 +1,138 @@
+"""
+Nick Depatie
+Assignment Processes
+This script houses a variety of functions used for processing data, such as analyzing list data
+    and mergesorting nested lists by urgency
+"""
+
+from datetime import datetime
+import math
+import numpy as np
+import time
+
+class process:
+
+    def month():
+        a=listtime[0]
+        if a == 'January' or a=='january' or a=='1' or a=='Jan':
+            listtime[0]='1'
+        if a == 'February' or a=='february' or a=='2' or a=='Feb':
+            listtime[0]='2'
+        if a == 'March' or a=='march' or a=='3' or a=='Mar':
+            listtime[0]='3'
+        if a == 'April' or a=='april' or a=='4' or a=='Apr':
+            listtime[0]='4'
+        if a == 'May' or a=='may' or a=='5' or a=='May':
+            listtime[0]='5'
+        if a == 'June' or a=='june' or a=='6' or a=='Jun':
+            listtime[0]='6'
+        if a == 'July' or a=='july' or a=='7' or a=='Jul':
+            listtime[0]='7'
+        if a == 'August' or a=='august' or a=='8' or a=='Aug':
+            listtime[0]='8'
+        if a == 'September' or a=='september' or a=='9' or a=='Sep':
+            listtime[0]='9'
+        if a == 'October' or a=='october' or a=='10' or a=='Oct':
+            listtime[0]='10'
+        if a == 'November' or a=='november' or a=='11' or a=='Nov':
+            listtime[0]='11'
+        if a == 'December' or a=='december' or a=='12' or a=='Dec':
+            listtime[0]='12'
+
+    def day():
+        a=listtime[1]
+        a=int(a)
+        if a<10:
+            listtime[1]=f'{a:02}'
+
+    def datetoint(inputtime):
+        global listtime
+        
+        listtime=list(inputtime.split(" "))
+        listtime.pop(0)
+        listtime.pop(2)
+        
+        process.month()
+        process.day()
+        
+        due_date=listtime[0]+listtime[1]
+        return due_date
+
+    def timeuntildeadline(due_date):
+        current_dt=datetime.now()
+        current_dt= int(current_dt.strftime("%m%d%H%M"))
+        deadline=due_date-current_dt
+        return deadline
+
+    def getsublist(superlist, n):
+        superlist=str(superlist)
+
+        unwantedchars="()'[]"+'"'
+        for character in unwantedchars:
+            superlist=superlist.replace(character,"")
+        sublist=superlist.split(",")
+
+        if n==0:
+            return sublist
+        elif n==1:
+            return sublist[4]
+        elif n==2:
+            superlist=superlist.replace(",","")
+            return superlist
+        elif n==3:
+            return sublist[1], sublist[0]
+        
+    def string_to_int(assignmentlist):
+        
+        for i in range (len(assignmentlist)):
+            assignments=process.getsublist(assignmentlist[i],0)
+            
+            deadline=assignments[2]
+            deadline=int(deadline.replace("'",""))
+            deadline=process.timeuntildeadline(deadline)
+            
+            priority=assignments[3]
+            priority=priority.replace("'","")
+            priority=int(priority)
+            
+            urgency=deadline-(1000*priority)
+
+            assignments.append(urgency)
+            assignments=str(assignments)
+            assignmentlist[i]=assignments
+    
+    def mergesort_list(list):
+        if len(list) > 1:
+
+            mid = len(list)//2
+            left = list[:mid]
+            right = list[mid:]
+
+            process.mergesort_list(left)
+            process.mergesort_list(right)
+
+            i=0
+            j=0
+            k=0
+
+            while i < len(left) and j < len(right):
+                if int(process.getsublist(left[i],1)) < int(process.getsublist(right[j],1)):
+                    list[k] = left[i]
+                    i += 1
+                    
+                else:
+                    list[k] = right[j]
+                    j += 1
+                    
+                k += 1
+                
+            # checking that no item was left behind
+            while i < len(left):
+                list[k] = left[i]
+                i += 1
+                k += 1
+
+            while j < len(right):
+                list[k] = right[j]
+                j += 1
+                k += 1
