+++ safeguardexample.py
@@ -1,0 +1,4 @@
+host="myHost",
+user="myUser",
+password="myPassword",
+database="myDatabase"
